phactory-api: Extract modules to files
