+++ crates/phactory/api/src/actions.rs
@@ -1,0 +1,7 @@
+pub const ACTION_GET_INFO: u8 = 2;
+
+pub const BIN_ACTION_START: u8 = 128;
+pub const BIN_ACTION_SYNC_PARA_HEADER: u8 = BIN_ACTION_START + 0;
+pub const BIN_ACTION_DISPATCH_BLOCK: u8 = BIN_ACTION_START + 1;
+pub const BIN_ACTION_SYNC_HEADER: u8 = BIN_ACTION_START + 2;
+pub const BIN_ACTION_SYNC_COMBINED_HEADERS: u8 = BIN_ACTION_START + 3;
